backend/requirements.txt
Django==5.1.5
psycopg2-binary==2.9.10
python-decouple==3.8
django-cors-headers==4.6.0
djangorestframework==3.15.2
Pillow==11.1.0
dj-database-url==2.1.0
gunicorn==23.0.0
whitenoise==6.8.2
djangorestframework-simplejwt==5.3.0
setuptools<81

# Additional packages for rental management
celery==5.3.6
redis==5.0.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
stripe==8.2.0
reportlab==4.0.8
openpyxl==3.1.2
django-filter==23.5
django-extensions==3.2.3
